Fix \MakeUppercase error in title
The error is:

```
Package beamerthememetropolis Warning: You need to compile with XeLaTeX or LuaL
aTeX to use the Fira fonts on input line 95.

))
(/usr/share/texlive/texmf-dist/tex/latex/beamer/beamerfontthemeprofessionalfont
s.sty))
! Undefined control sequence.
\MakeUppercase ...ppercaseUnsupportedInPdfStrings
```

The issue stems from https://tex.stackexchange.com/a/199385

The reason, if I understand vaguely, is that the title is parsed before the command is properly defined.